Initial commit - stupid simple interpreter
I have a running interpreter for a simple extensible
stack-based postfix calculator.

My calculator is based on forth and is extensible in that you
can define new words (which is basically how forth has functions)
but it's not turing complete because there are no control
structures.

I have dreams of either making it a proper forth - probably based
on easyforth - or making a forth-ish thing that has different
semantics - probably based on STCK.

I included some detailed notes on the implementations, and on my
next steps (which are to read rather than code).